backend/requirements.txt
boto3>=1.34,<2
base58>=2.1,<3
PyNaCl>=1.5,<2
